Moved OTel stuff to io.opentelemetry.rum.internal
