+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/OpenTelemetryRumBuilder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.splunk.rum.internal.opentelemetry;
+package io.opentelemetry.rum.internal;
 
 import android.app.Application;
 import io.opentelemetry.sdk.metrics.SdkMeterProviderBuilder;
